test: create constants module
