+++ test/validate.spec.js
@@ -4,6 +4,7 @@ const Ajv = require('ajv');
 const addFormats = require('ajv-formats');
 const { readFileSync } = require('fs');
 const { resolve } = require('path');
+const { languages, translationProperties } = require('./constants');
 
 const createAjvInstance = () => {
   const ajv = new Ajv({
@@ -146,100 +147,6 @@ describe('Validate', function () {
       ],
     },
   ];
-
-  const languages = ['DE', 'EN', 'FR', 'PL'];
-  const translationProperties = {
-    certificateFields: [
-      'A01',
-      'A02',
-      'A03',
-      'A04',
-      'A05',
-      'A06',
-      'A06.1',
-      'A06.2',
-      'A06.3',
-      'A07',
-      'A08',
-      'A09',
-      'A97',
-      'A98',
-      'A99',
-      'B01',
-      'B02',
-      'B03',
-      'B04',
-      'B05',
-      'B06',
-      'B07',
-      'B08',
-      'B09',
-      'B10',
-      'B11',
-      'B12',
-      'B13',
-      'C00',
-      'C01',
-      'C02',
-      'C03',
-      'C10',
-      'C11',
-      'C12',
-      'C13',
-      'C30',
-      'C31',
-      'C32',
-      'C40',
-      'C41',
-      'C42',
-      'C43',
-      'C70',
-      'D01',
-      'Z01',
-      'Z02',
-      'Z03',
-      'Z04',
-      'Z04.NotifiedBody',
-      'Z04.DoCYear',
-      'Z04.DoCNumber',
-    ],
-    certificateGroups: ['CommercialTransaction', 'ProductDescription', 'Inspection', 'OtherTests', 'Validation'],
-    otherFields: [
-      'AdditionalInformation',
-      'SupplementaryInformation',
-      'TensileTest',
-      'HardnessTest',
-      'ChemicalComposition',
-      'NotchedBarImpactTest',
-      'OtherMechanicalTests',
-      'NonDestructiveTests',
-      'OtherProductTests',
-      'ProductNorm',
-      'MaterialNorm',
-      'MassNorm',
-      'SteelDesignation',
-      'Form',
-      'Tube',
-      'QuadraticTube',
-      'RectangularTube',
-      'Pipe',
-      'RectangularPipe',
-      'Coil',
-      'RoundBar',
-      'HexagonalBar',
-      'FlatBar',
-      'Validation',
-      'Other',
-      'Width',
-      'Thickness',
-      'OuterDiameter',
-      'Diameter',
-      'WallThickness',
-      'Height',
-      'SideLength',
-      'Unit',
-    ],
-  };
 
   it('should validate schema', () => {
     const validateSchema = createAjvInstance().compile(localSchema);
